requirements.txt
d3blocks==1.6.1
matplotlib==3.10.7
numpy==2.3.4
pandas==2.3.3
Pillow==12.0.0
playwright==1.55.0
PyYAML==6.0.3
redshift_connector==2.1.8
seaborn==0.13.2
